# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_handler.cancelarNotaRequest:
    properties:
      justificativa:
        type: string
    type: object
...
  internal_handler.fecharPagamentoRequest:
    properties:
      comanda_ids:
        items:
          type: string
        type: array
      pagamentos:
        items:
          $ref: '#/definitions/internal_handler.pagamentoParcialRequest'
        type: array
    type: object
  internal_handler.fecharPagamentoResponse:
    properties:
      payment_ids:
        items:
          type: string
        type: array
    type: object
...
  internal_handler.loginRequest:
    properties:
      login:
        type: string
      senha:
        type: string
    type: object
  internal_handler.loginResponse:
    properties:
      token:
        type: string
    type: object
...
  internal_handler.pagamentoParcialRequest:
    properties:
      metodo:
        type: string
      valor:
        type: number
    type: object
...
info:
  contact: {}
  description: Sistema de comandas multi-tenant (churrascaria como primeira instância).
    Ver CLAUDE.md e docs/merka-planejamento.md para o contexto completo do domínio.
  title: Merka API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 'Valida login/senha e devolve um JWT contendo user_id, tenant_id
        e role_id, usado nas demais rotas via header Authorization: Bearer <token>.'
      parameters:
      - description: Credenciais
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.loginResponse'
        "400":
          description: corpo da requisição inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: login ou senha inválidos
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Autenticar usuário
      tags:
      - auth
...
  /pagamentos:
    post:
      consumes:
      - application/json
      description: Soma o total ativo (itens + pesos, excluindo removidos/estornados)
        das comandas informadas — que podem ser N comandas de uma mesma mesa — e confere
        contra a soma dos pagamentos parciais informados (suporta pagamento misto).
        Se bater, grava um payment por método, liga todas as comandas via payment_comandas
        e marca as comandas como "paga". Não emite nota fiscal ainda (ver TODO no
        usecase para US-14/emissão de NFC-e).
      parameters:
      - description: Comandas a fechar e pagamentos parciais
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.fecharPagamentoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.fecharPagamentoResponse'
        "400":
          description: corpo inválido, método inválido ou soma dos pagamentos não
            bate com o total
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: token ausente, inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: alguma comanda não encontrada
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Fechar pagamento de uma ou mais comandas (US-13 + US-14)
      tags:
      - pagamentos
  /pagamentos/{id}/cancelar-nota:
    post:
      consumes:
      - application/json
      description: Cancela a NFC-e vinculada ao pagamento informado, junto à SEFAZ
        (integração direta, ver CLAUDE.md). Só permitido dentro do prazo regulamentar
        a partir da emissão — passado esse prazo a SEFAZ rejeita o evento, sem reenvio
        automático. Restrito a Caixa/Gestor/Admin Super.
      parameters:
      - description: id do payment
        in: path
        name: id
        required: true
        type: string
      - description: Justificativa do cancelamento (15-255 caracteres, exigido pela
          SEFAZ)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.cancelarNotaRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: justificativa inválida, nota não emitida, já cancelada, ou
            prazo de cancelamento expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: token ausente, inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: nenhuma nota fiscal encontrada pra esse pagamento
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancelar NFC-e já emitida, dentro do prazo (US-22)
      tags:
      - pagamentos
...
securityDefinitions:
  BearerAuth:
    description: Informe "Bearer <token>" (token obtido em POST /auth/login)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"